test(flare): relocate tests to top-level tests/ directory and consolidate
